refactor(domain)!: replace type `Int` with `Long` for IDs

+++ app/src/main/java/tech/salvatore/livro_android_kotlin_paulo_salvatore/model/source/local/db/entity/CreatureEntity.kt
@@ -13,11 +13,11 @@ import androidx.room.PrimaryKey
 )
 data class CreatureEntity(
     @PrimaryKey
-    val number: Int,
+    val number: Long,
 
     val name: String,
 
     val imageUrl: String,
 
-    val evolveToNumber: Int?,
+    val evolveToNumber: Long?,
 )
